removed home component and replaced the main page with it
